% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_Ipaper.R
\name{check_file}
\alias{check_file}
\title{check_file
Check file whether exist. If not, then give a notification.}
\usage{
check_file(file, duration = 10)
}
\description{
check_file
Check file whether exist. If not, then give a notification.
}
