% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.shiny-package.R
\docType{package}
\name{phenofit.shiny-package}
\alias{phenofit.shiny}
\alias{phenofit.shiny-package}
\title{phenofit.shiny: Shiny app of phenofit}
\description{
The merits of 'TIMESAT' and 'phenopix' are adopted. Besides, a simple and 
    growing season dividing method and a practical snow elimination method 
    based on Whittaker were proposed. 7 curve fitting methods and 4 phenology 
    extraction methods were provided. Parameters boundary are considered for 
    every curve fitting methods according to their ecological meaning. 
    And 'optimx' is used to select best optimization method for different 
    curve fitting methods.
    Reference:
    Dongdong Kong, R package: A state-of-the-art Vegetation Phenology extraction package, 
    phenofit version 0.2.3, <https://github.com/kongdd/phenofit>;
    Zhang, Q., Kong, D., Shi, P., Singh, V.P., Sun, P., 2018. Vegetation phenology 
    on the Qinghai-Tibetan Plateau and its response to climate change (1982–2013). 
    Agric. For. Meteorol. 248, 408–417. <doi:10.1016/j.agrformet.2017.10.026>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kongdd/phenofit}
  \item Report bugs at \url{https://github.com/kongdd/phenofit/issues}
}

}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
