% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_site.R
\name{phenofit_site}
\alias{phenofit_site}
\title{Extract Vegetation Phenology at site scale}
\usage{
phenofit_site(y, t, w, QC_flag, nptperyear = 36, prefix = "",
  titlestr = NULL, show = FALSE)
}
\arguments{
\item{y}{Numeric vector, vegetation index time-series}

\item{t}{Numeric vector, \code{Date} variable}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{QC_flag}{Factor (optional) returned by \code{qcFUN}, levels should be
in the range of \code{c("snow", "cloud", "shadow", "aerosol", "marginal", "good")}, others will be categoried into \code{others}. \code{QC_flag} is
used for visualization in \code{\link[=get_pheno]{get_pheno()}} and \code{\link[=plot_phenofit]{plot_phenofit()}}.}

\item{nptperyear}{Integer, number of images per year.}
}
\description{
Extract Vegetation Phenology at site scale
}
