% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{resample_spatialPixel}
\alias{resample_spatialPixel}
\title{resample_spatialPixel}
\usage{
resample_spatialPixel(df, r, r.new, I_grid, I_grid.new = NULL,
  verbose = TRUE)
}
\arguments{
\item{df}{nrow(df) should be equal to nrow(r)}

\item{r}{raster}

\item{I_grid}{Index of selected pixel of raster (from raster::raster)}
}
\description{
resample_spatialPixel
}
